Address comments - Update requests/insights using useEffect, move group actions on user to api folder, remove unnecesary asyncs
